GameSceneParser.swiftを解体
BMSファイルのパースとノーツオブジェクトの生成を分離。
MusicクラスのイニシャライザでBMSファイルからノーツの中間表現を作成し、
MusicクラスのgenerateNotesメソッドでノーツオブジェクトを生成するように変更
