+++ 音ゲー（仮）/Note.swift
@@ -13,12 +13,13 @@ import SpriteKit
 class Tap: Note {
     
     let isLarge: Bool                   // 大ノーツかどうか
-    var appearTime: TimeInterval = 0    // 演奏開始から水平線を超えるまでの時間。これ以降にposの計算&更新を行う。
+    let appearTime: TimeInterval        // 演奏開始から水平線を超えるまでの時間。これ以降にposの計算&更新を行う。
     
-    init(beatPos beat: Double, laneIndex: Int, speedRatio: Double, isLarge: Bool) {
-        self.isLarge = isLarge
-        super.init(beatPos: beat, laneIndex: laneIndex, speedRatio: speedRatio)
-        
+    init(rawNote: RawNote, speed: CGFloat, appearTime: TimeInterval) {
+        self.isLarge = rawNote.isLarge
+        self.appearTime = appearTime
+        super.init(beatPos: rawNote.beat, laneIndex: rawNote.laneIndex, speed: speed)
+
         // imageのインスタンス(白円or黄円)を作成
         self.image = SKShapeNode(circleOfRadius: Note.initialSize / 2)
         image.fillColor = isLarge ? UIColor.yellow : UIColor.white
@@ -60,11 +61,12 @@ class Tap: Note {
 /// 呼び出し時にまだparfectの時間でない場合(before)について、後にparfect判定を行うかもしれないので、時間とUItouch情報を該当LaneインスタンスのstoredFlickJudgeに、レーン情報を該当GSTouchインスタンスのstoredFlickJudgeLaneIndexに格納し、後にこの情報をもとにGameSceneTouchesファイル内に記述されているGameScene.storedFlickJudge関数にて判定を行う。この呼出は情報が残っているときにのみ行われ、該当ノーツの判定後に各情報格納場所にnilが入る。storedFlickJudgeの呼び出しタイミングはtouchesMoved呼び出し時にレーンから指が外れた時、touchesEnded呼び出し時、これ以上待ってもより良い判定が来なくなる時（ノーツの正確なタイミングの時間についてtimeLag予定時間(>0)と対象な時間）である。
 class Flick: Note {
     
-    var appearTime: TimeInterval = 0        // 演奏開始から水平線を超えるまでの時間。これ以降にposの計算&更新を行う。
+    let appearTime: TimeInterval        // 演奏開始から水平線を超えるまでの時間。これ以降にposの計算&更新を行う。
     
-    override init(beatPos beat: Double, laneIndex: Int, speedRatio: Double) {
-        super.init(beatPos: beat, laneIndex: laneIndex, speedRatio: speedRatio)
-        
+    init(rawNote: RawNote, speed: CGFloat, appearTime: TimeInterval) {
+        self.appearTime = appearTime
+        super.init(beatPos: rawNote.beat, laneIndex: rawNote.laneIndex, speed: speed)
+
         // imageのインスタンス(マゼンタ三角形)を作成
         let length = Note.initialSize / 2   // 三角形一辺の長さの半分
         // 始点から終点までの４点を指定(2点を一致させ三角形に).
@@ -114,12 +116,13 @@ class TapStart: Note {
     var next = Note()                                               // 次のノーツ（仮のインスタンス）
     var longImages = (long: SKShapeNode(), circle: SKShapeNode())   // このノーツを始点とする緑太線の画像と、判定線上に残る緑楕円(将来的にはimageに格納？)
     let isLarge: Bool                                               // 大ノーツかどうか
-    var appearTime: TimeInterval = 0                                // 演奏開始から水平線を超えるまでの時間。これ以降にposの計算&更新を行う。
+    let appearTime: TimeInterval                                // 演奏開始から水平線を超えるまでの時間。これ以降にposの計算&更新を行う。
     
-    init(beatPos beat: Double, laneIndex: Int, speedRatio: Double, isLarge: Bool) {
-        self.isLarge = isLarge
-        super.init(beatPos: beat, laneIndex: laneIndex, speedRatio: speedRatio)
-        
+    init(rawNote: RawNote, speed: CGFloat, appearTime: TimeInterval) {
+        self.isLarge = rawNote.isLarge
+        self.appearTime = appearTime
+        super.init(beatPos: rawNote.beat, laneIndex: rawNote.laneIndex, speed: speed)
+
         // imageのインスタンス(緑円or黄円)を作成
         image = SKShapeNode(circleOfRadius: Note.initialSize / 2)
         image.fillColor = isLarge ? UIColor.yellow : UIColor.green
@@ -260,7 +263,7 @@ class Middle: Note {
     
     var next = Note()                                               // 次のノーツ（仮のインスタンス）
     var longImages = (long: SKShapeNode(), circle: SKShapeNode())   // このノーツを始点とする緑太線の画像と、判定線上に残る緑楕円(将来的にはimageに格納？)
-    var appearTime: TimeInterval = 0                                // 演奏開始から水平線を超えるまでの時間。これ以降にposの計算&更新を行う。
+    let appearTime: TimeInterval                                    // 演奏開始から水平線を超えるまでの時間。これ以降にposの計算&更新を行う。
     override var position: CGPoint {                                // positionを左端ではなく線の中点にするためオーバーライド
         get {
             return CGPoint(x: image.position.x + size / 2, y: image.position.y)
@@ -270,9 +273,10 @@ class Middle: Note {
         }
     }
     
-    override init(beatPos beat: Double, laneIndex: Int, speedRatio: Double) {
-        super.init(beatPos: beat, laneIndex: laneIndex, speedRatio: speedRatio)
-        
+    init(rawNote: RawNote, speed: CGFloat, appearTime: TimeInterval) {
+        self.appearTime = appearTime
+        super.init(beatPos: rawNote.beat, laneIndex: rawNote.laneIndex, speed: speed)
+
         self.isJudgeable = false
         
         // imageのインスタンス(緑線分)を作成
@@ -409,13 +413,13 @@ class Middle: Note {
 /// touchesEnded呼び出し時に判定を行う。
 class TapEnd: Note {
     
-    unowned var start = Note()  // 循環参照防止の為unowned参照にする
+//    unowned var start = Note()  // 循環参照防止の為unowned参照にする
     let isLarge: Bool           // 大ノーツかどうか
     
-    init(beatPos beat: Double, laneIndex: Int, speedRatio: Double, isLarge: Bool) {
-        self.isLarge = isLarge
-        super.init(beatPos: beat, laneIndex: laneIndex, speedRatio: speedRatio)
-        
+    init(rawNote: RawNote, speed: CGFloat) {
+        self.isLarge = rawNote.isLarge
+        super.init(beatPos: rawNote.beat, laneIndex: rawNote.laneIndex, speed: speed)
+
         self.isJudgeable = false
         
         // imageのインスタンス(緑円or黄円)を作成
@@ -457,10 +461,10 @@ class TapEnd: Note {
 /// 呼び出し時にまだparfectの時間でない場合(before)について、後にparfect判定を行うかもしれないので、時間とUItouch情報を該当LaneインスタンスのstoredFlickJudgeに、レーン情報を該当GSTouchインスタンスのstoredFlickJudgeLaneIndexに格納し、後にこの情報をもとにGameSceneTouchesファイル内に記述されているGameScene.storedFlickJudge関数にて判定を行う。この呼出は情報が残っているときにのみ行われ、該当ノーツの判定後に各情報格納場所にnilが入る。storedFlickJudgeの呼び出しタイミングはtouchesMoved呼び出し時にレーンから指が外れた時、touchesEnded呼び出し時、これ以上待ってもより良い判定が来なくなる時（ノーツの正確なタイミングの時間についてtimeLag予定時間(>0)と対象な時間）である。
 class FlickEnd: Note {
     
-    unowned var start = Note()
-    
-    override init(beatPos beat: Double, laneIndex: Int, speedRatio: Double) {
-        super.init(beatPos: beat, laneIndex: laneIndex, speedRatio: speedRatio)
+//    unowned var start = Note()
+
+    init(rawNote: RawNote, speed: CGFloat) {
+        super.init(beatPos: rawNote.beat, laneIndex: rawNote.laneIndex, speed: speed)
         
         self.isJudgeable = false
         
@@ -520,108 +524,105 @@ class Note {
         }
     }
     fileprivate var positionOnLane: CGFloat = 0.0           // ノーツのレーン上の座標(判定線を0、奥を正の向きとする)
-    private let baseSpeed: CGFloat                          // ノーツスピード。実際のスピードはこの値とBPMによって決定される
+    fileprivate let speed: CGFloat                          // ノーツスピード。ユーザー設定とBPMによって決定される
     static var scale: CGFloat = 1.0                         // ノーツの幅の倍率(ノーツごとの差異はなく、メモリ領域削減のためstatic)。settingのほか、レーン数を踏まえて値を返す
-    fileprivate static let longScale: CGFloat = 0.8                // ノーツの幅に対するlongの幅の倍率
+    static var BPMs: [(bpm: Double, startPos: Double)] = [] // GameSceneのBPMsと同じもの
+    fileprivate static let longScale: CGFloat = 0.8         // ノーツの幅に対するlongの幅の倍率
     fileprivate static let initialSize = CGFloat(100)       // ノーツの初期サイズ。ノーツ大きさはscaleで調節するのでどんな値でもよい
-    private static var majorBPM: Double = 0.0               // 楽曲の基本BPM。BPM配列の中から最も持続時間が長いもの。
-    private static var BPMs: [(bpm: Double, startPos: Double)] = []
-    
-    
-    
+//    private static var majorBPM: Double = 0.0               // 楽曲の基本BPM。BPM配列の中から最も持続時間が長いもの。
+
+
     /// Noteのイニシャライザ
     ///
     /// - Parameters:
     ///   - beat: 拍
     ///   - laneIndex: laneのindex
-    ///   - speedRatio: bmsの指定を反映した、各ノーツ固有のスピード倍率
-    ///   - setting: 設定
-    init(beatPos beat: Double, laneIndex: Int, speedRatio: Double) {
+    ///   - speed: 3Dレーン上のノーツ秒速
+    init(beatPos beat: Double, laneIndex: Int, speed: CGFloat) {
         self.beat = beat
         self.laneIndex = laneIndex
-        self.baseSpeed = 1350 * CGFloat(speedRatio)
+        self.speed = speed
     }
     init() {
         self.beat = 0
         self.laneIndex = 0
-        self.baseSpeed = 1350
+        self.speed = 1350
     }
     
     deinit {
         self.image.removeFromParent()
     }
     
-    /// クラスプロパティとappearTimeを設定.必ずパース後に実行すること.
-    static func initialize(_ duration: TimeInterval, _ notes: [Note], _ music: Music, _ setting: Setting) {
-        
-        guard !music.BPMs.isEmpty else {
-            print("空のBPM配列")
-            return
-        }
-        
-        var BPMIntervals: [(bpm: Double, interval: TimeInterval)] = []
-        var timeSum: TimeInterval = 0
-        var i = 0
-        while true {
-            if i + 1 < music.BPMs.count {
-                let interval = TimeInterval((music.BPMs[i + 1].startPos - music.BPMs[i].startPos) / (music.BPMs[i].bpm/60))
-                BPMIntervals.append((music.BPMs[i].bpm, interval))
-                timeSum += interval
-            } else {
-                let interval = duration - timeSum
-                BPMIntervals.append((music.BPMs[i].bpm, interval))
-                break
-            }
-            i += 1
-        }
-        Note.majorBPM = BPMIntervals.max { $0.interval < $1.interval }!.bpm
-        Note.BPMs = music.BPMs
-        
-        if !setting.isFitSizeToLane {
-            Note.scale = CGFloat(setting.scaleRatio/7*Double(music.laneNum))
-        } else {
-            Note.scale = CGFloat(setting.scaleRatio)
-        }
-        
-        // appearTimeの設定(end系以外)
-        for note in notes {
-            switch note {
-            case is Tap:
-                let tap = note as! Tap
-                tap.appearTime = getAppearTime(note)
-            case is Flick:
-                let flick = note as! Flick
-                flick.appearTime = getAppearTime(note)
-            case is TapStart:
-                let tapStart = note as! TapStart
-                tapStart.appearTime = getAppearTime(note)
-            case is Middle:
-                let middle = note as! Middle
-                middle.appearTime = getAppearTime(note)
-            default:
-                break
-            }
-        }
-    }
-    
-    /// ノーツが画面上に現れる時刻を返す(updateするかの判定に使用)
-    private static func getAppearTime(_ note: Note) -> TimeInterval {
-        
-        var judgeTime: TimeInterval = 0.0   // 判定線所雨に乗る時刻
-        var i = 0
-        while i + 1 < Note.BPMs.count && Note.BPMs[i + 1].startPos < note.beat {
-            judgeTime += (Note.BPMs[i + 1].startPos - Note.BPMs[i].startPos) / (Note.BPMs[i].bpm/60)
-            
-            i += 1
-        }
-        judgeTime += (note.beat - Note.BPMs[i].startPos) / (Note.BPMs[i].bpm/60)
-        
-        let speed = note.baseSpeed * CGFloat(Note.BPMs[i].bpm / Note.majorBPM)
-        let appearTime = judgeTime - TimeInterval(Dimensions.laneLength / speed)   // judgeTime - レーン端から端までかかる時間
-        
-        return appearTime
-    }
-    
+//    /// クラスプロパティとappearTimeを設定.必ずパース後に実行すること.
+//    static func initialize(_ duration: TimeInterval, _ notes: [Note], _ music: Music, _ setting: Setting) {
+//
+//        guard !music.BPMs.isEmpty else {
+//            print("空のBPM配列")
+//            return
+//        }
+//
+//        var BPMIntervals: [(bpm: Double, interval: TimeInterval)] = []
+//        var timeSum: TimeInterval = 0
+//        var i = 0
+//        while true {
+//            if i + 1 < music.BPMs.count {
+//                let interval = TimeInterval((music.BPMs[i + 1].startPos - music.BPMs[i].startPos) / (music.BPMs[i].bpm/60))
+//                BPMIntervals.append((music.BPMs[i].bpm, interval))
+//                timeSum += interval
+//            } else {
+//                let interval = duration - timeSum
+//                BPMIntervals.append((music.BPMs[i].bpm, interval))
+//                break
+//            }
+//            i += 1
+//        }
+//        Note.majorBPM = BPMIntervals.max { $0.interval < $1.interval }!.bpm
+//        Note.BPMs = music.BPMs
+//
+//        if !setting.isFitSizeToLane {
+//            Note.scale = CGFloat(setting.scaleRatio/7*Double(music.laneNum))
+//        } else {
+//            Note.scale = CGFloat(setting.scaleRatio)
+//        }
+//
+//        // appearTimeの設定(end系以外)
+//        for note in notes {
+//            switch note {
+//            case is Tap:
+//                let tap = note as! Tap
+//                tap.appearTime = getAppearTime(note)
+//            case is Flick:
+//                let flick = note as! Flick
+//                flick.appearTime = getAppearTime(note)
+//            case is TapStart:
+//                let tapStart = note as! TapStart
+//                tapStart.appearTime = getAppearTime(note)
+//            case is Middle:
+//                let middle = note as! Middle
+//                middle.appearTime = getAppearTime(note)
+//            default:
+//                break
+//            }
+//        }
+//    }
+
+//    /// ノーツが画面上に現れる時刻を返す(updateするかの判定に使用)
+//    fileprivate func getAppearTime() -> TimeInterval {
+//
+//        var judgeTime: TimeInterval = 0.0   // 判定線上に乗る時刻
+//        var i = 0
+//        while i + 1 < Note.BPMs.count && Note.BPMs[i + 1].startPos < beat {
+//            judgeTime += (Note.BPMs[i + 1].startPos - Note.BPMs[i].startPos) / (Note.BPMs[i].bpm/60)
+//
+//            i += 1
+//        }
+//        judgeTime += (beat - Note.BPMs[i].startPos) / (Note.BPMs[i].bpm/60)
+//
+//        let appearTime = judgeTime - TimeInterval(Dimensions.laneLength / speed)   // judgeTime - レーン端から端までかかる時間
+//
+//        return appearTime
+//    }
+
     /// ノーツの表示状態の更新、毎フレーム呼ばれる
     /// 各派生クラスでオーバーライドされる
     ///
@@ -643,7 +644,7 @@ class Note {
         remainingTime += (beat - Note.BPMs[i].startPos) / (Note.BPMs[i].bpm/60)
         remainingTime -= passedTime
         
-        self.positionOnLane = CGFloat(remainingTime) * baseSpeed * CGFloat(Note.BPMs[i].bpm / Note.majorBPM)    // 判定線からの水平距離x
+        self.positionOnLane = CGFloat(remainingTime) * speed    // 判定線からの水平距離x
     }
     
     /// ノーツの座標を設定
